+++ python3/ltk/actions/config_action.py
@@ -377,7 +377,7 @@ class ConfigAction(Action):
             log_info = 'Removing all target locales'
             self.update_config_file('watch_locales', '', self.conf_parser, self.config_file_name, log_info)
         else:
-            target_locales = get_valid_locales(self.api,locales)
+            target_locales = get_valid_locales(self.api,locales,'added')
             target_locales_str = ','.join(target for target in target_locales)
             if len(target_locales_str) > 0:
                 log_info = 'Set target locales to {0}'.format(target_locales_str)