+++ python3/ltk/apicalls.py
@@ -294,6 +294,16 @@ class ApiCalls:
             self.handleError()
         return r
 
+    def document_translation_locale_status(self, document_id, locale):
+        """ gets the status of a specific target for a document translation """
+        try:
+            uri = (API_URI['document_translation_locale'] % locals())
+            r = requests.get(self.host + uri, headers=self.headers)
+            log_api('GET', uri, r)
+        except requests.exceptions.ConnectionError:
+            self.handleError()
+        return r
+
     def document_content(self, document_id, locale_code, auto_format, xliff=False, finalized_file='off'):
         """ downloads the translated document """
         headers = self.headers
@@ -348,11 +358,31 @@ class ApiCalls:
             self.handleError()
         return r
 
+    def document_cancel_target(self, document_id, locale):
+        try:
+            uri = (API_URI['document_cancel_locale'] % locals())
+            payload = {'id': document_id, 'locale': locale, 'cancelled_reason': 'CANCELLED_BY_AUTHOR', 'mark_invoiceable': 'true'}
+            r = requests.post(self.host + uri, headers=self.headers, data=payload)
+            log_api('POST', uri, r)
+        except requests.exceptions.ConnectionError:
+            self.handleError()
+        return r
+
     def document_delete(self, document_id):
         try:
             uri = (API_URI['document_id'] % locals())
             r = requests.delete(self.host + uri, headers=self.headers)
             log_api('DELETE', uri, r)
+        except requests.exceptions.ConnectionError:
+            self.handleError()
+        return r
+
+    def document_cancel(self, document_id):
+        try:
+            uri = (API_URI['document_cancel'] % locals())
+            payload = {'id': document_id, 'cancelled_reason': 'CANCELLED_BY_AUTHOR'}
+            r = requests.post(self.host + uri, headers=self.headers, data=payload)
+            log_api('POST', uri, r)
         except requests.exceptions.ConnectionError:
             self.handleError()
         return r