+++ tests/test_actions/test_download.py
@@ -28,7 +28,7 @@ class TestDownload(unittest.TestCase):
         self.locales = ['ja-JP', 'zh-CN']
         self.action = DownloadAction(os.getcwd())
         self.clean_action = CleanAction(os.getcwd())
-        self.request_action = RequestAction(os.getcwd(), None, None, self.locales, False, None, None)
+        self.request_action = RequestAction(os.getcwd(), None, None, self.locales, False, False, None, None)
         self.pull_action = PullAction(os.getcwd(), self.action)
         self.clean_action.clean_action(False, False, None)
         self.files = ['sample.txt', 'sample1.txt']
@@ -44,7 +44,7 @@ class TestDownload(unittest.TestCase):
     def tearDown(self):
         self.rm_action = RmAction(os.getcwd())
         for curr_file in self.files:
-            self.rm_action.rm_action([curr_file], force=True)
+            self.rm_action.rm_action([curr_file], remote=True, force=True)
         self.clean_action.clean_action(False, False, None)
         for dl_file in self.downloaded_files:
             if os.path.exists(dl_file):