+++ tests/test_actions/test_list.py
@@ -47,7 +47,7 @@ class TestList(unittest.TestCase):
             sys.stdout = sys.__stdout__
 
         for fn in files:
-            self.rm_action.rm_action(fn, force=True)
+            self.rm_action.rm_action(fn, remote=True, force=True)
         self.clean_action.clean_action(False, False, None)
 
     def test_list_docs_none(self):
@@ -113,3 +113,95 @@ class TestList(unittest.TestCase):
         # list
         # check
         pass
+    
+    def test_list_doc_remote(self):
+        files = ['sample.txt', 'sample1.txt', 'sample2.txt']
+        file_paths = []
+        for fn in files:
+            file_paths.append(create_txt_file(fn))
+        self.add_action.add_action(['sample*.txt'], overwrite=True)
+        doc_ids = self.action.doc_manager.get_doc_ids()
+        for doc_id in doc_ids:
+            assert poll_doc(self.action, doc_id)
+        try:
+            out = StringIO()
+            sys.stdout = out
+            self.action.list_remote()
+            info = out.getvalue()
+            for doc_id in doc_ids:
+                assert doc_id in info
+        finally:
+            sys.stdout = sys.__stdout__
+
+        for fn in files:
+            self.rm_action.rm_action(fn, remote=True, force=True)
+        self.clean_action.clean_action(False, False, None)
+
+    #can't test a remote list of none because there is no guarantee that the project used in the config file for the test is empty, and we don't want to clear the project remotely in case it is one that has stuff unrelated to testing that needs to stay
+    #def test_list_docs_remote_none(self):
+    #    try:
+    #        out = StringIO()
+    #        sys.stdout = out
+    #        self.action.list_remote()
+    #        info = out.getvalue()
+    #        assert 'No documents to report' in info
+    #    finally:
+    #        sys.stdout = sys.__stdout__
+
+    def test_target_download_folder(self):
+        files = ['sample.txt', 'sample1.txt', 'sample2.txt']
+        file_paths = []
+        for fn in files:
+            file_paths.append(create_txt_file(fn))
+        directory = os.path.join(os.getcwd(), 'test_dir')
+        create_directory(directory)
+        self.add_action.add_action(['sample.txt'], overwrite=True, download_folder='test_dir')
+        self.add_action.add_action(['sample1.txt'], overwrite=True, download_folder='.')
+        self.add_action.add_action(['sample2.txt'], overwrite=True, download_folder='')
+        doc_ids = self.action.doc_manager.get_doc_ids()
+        for doc_id in doc_ids:
+            assert poll_doc(self.action, doc_id)
+        try:
+            out = StringIO()
+            sys.stdout = out
+            self.action.list_action(hide_docs=False, title=False, show_dests=True)
+            info = out.getvalue()
+            import re
+            match1 = re.search('\nsample.txt\s*[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}\s*test_dir\s*\n', info)
+            assert match1
+            match2 = re.search('\nsample1.txt\s*[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}\s*\.\s*\n', info)
+            assert match2
+            match3 = re.search('\nsample2.txt\s*[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}\s*\n', info)
+            assert match3
+        finally:
+            sys.stdout = sys.__stdout__
+
+        for fn in files:
+            self.rm_action.rm_action(fn, remote=True, force=True)
+        self.clean_action.clean_action(False, False, None)
+        delete_directory(directory)
+
+    def test_list_cancelled(self):
+        files = ['sample.txt', 'sample1.txt', 'sample2.txt']
+        file_paths = []
+        for fn in files:
+            file_paths.append(create_txt_file(fn))
+        self.add_action.add_action(['sample*.txt'], overwrite=True)
+        doc_ids = self.action.doc_manager.get_doc_ids()
+        for doc_id in doc_ids:
+            assert poll_doc(self.action, doc_id)
+        self.action.api.document_cancel(doc_ids[0])
+        assert poll_rm(self.action, doc_ids[0], cancelled=True)
+        try:
+            out = StringIO()
+            sys.stdout = out
+            self.action.list_ids(False)
+            info = out.getvalue()
+            for doc_id in doc_ids:
+                assert doc_id in info
+        finally:
+            sys.stdout = sys.__stdout__
+
+        for fn in files:
+            self.rm_action.rm_action(fn, remote=True, force=True)
+        self.clean_action.clean_action(False, False, None)